simplify pickers, move css to app layer
