+++ assets/src/infrastructure/ui/inputs/dateTime/Datepicker/index.tsx
@@ -1,5 +1,4 @@
 import React, { useCallback, useState } from 'react';
-import classNames from 'classnames';
 import { default as ReactDatepicker } from 'react-date-picker';
 
 import { DatePickerProps } from '../types';
@@ -7,19 +6,15 @@ import { CalendarOutlined, CloseOutlined } from '@appDisplay/icons/svgs';
 import useConfig from '@appServices/config/useConfig';
 import { convertWordPressDateFormat } from '../utilities';
 
-import './style.scss';
+import '../style.scss';
 
-const Datepicker: React.FC<DatePickerProps> = ({ value, onChange, onChangeValue, ...props }) => {
+const DatePicker: React.FC<DatePickerProps> = ({ value, onChange, onChangeValue, ...props }) => {
 	const [date, setDate] = useState(value);
 	const {
 		dateTimeFormats: { dateFormat },
 		locale: { user },
 	} = useConfig();
-	const convertedDateFormat = convertWordPressDateFormat(dateFormat);
-	const className = classNames(
-		'ee-input-base-wrapper ee-date-picker',
-		props.className
-	);
+
 	const onChangeHandler: DatePickerProps['onChange'] = useCallback(
 		(newDate) => {
 			setDate(newDate);
@@ -31,27 +26,27 @@ const Datepicker: React.FC<DatePickerProps> = ({ value, onChange, onChangeValue,
 				onChangeValue(newDate as Date);
 			}
 
-			if ( typeof onChange === 'function' ) {
+			if (typeof onChange === 'function') {
 				onChange(newDate);
 			}
 		},
-		[ onChange, onChangeValue, ]
+		[onChange, onChangeValue]
 	);
 
+	// convert date format to accepatble values for react-date-picker
+	const convertedDateFormat = convertWordPressDateFormat(dateFormat);
+
 	return (
-		<div className={className}>
-			<ReactDatepicker
-				format={convertedDateFormat}
-				{...props}
-				calendarIcon={<CalendarOutlined />}
-				clearIcon={<CloseOutlined />}
-				locale={user}
-				onChange={onChangeHandler}
-				required
-				value={date}
-			/>
-		</div>
+		<ReactDatepicker
+			format={convertedDateFormat}
+			{...props}
+			calendarIcon={<CalendarOutlined />}
+			clearIcon={<CloseOutlined />}
+			locale={user}
+			onChange={onChangeHandler}
+			value={date}
+		/>
 	);
 };
 
-export default Datepicker;
+export default DatePicker;
