--- assets/src/infrastructure/ui/inputs/dateTime/Timepicker/style.scss
@@ -1,194 +1,0 @@
-/** -------------------------------------------------------------------------
- * Event Espresso Date Picker Stylesheet
- * © 2020 Event Espresso
- * ------------------------------------------------------------------------- */
-
-.ee-time-picker-active {
-	border-color: var(--ee-color-primary);
-}
-
-.ee-time-picker-focus {
-	border-color: var(--ee-color-primary);
-	/* Visible in Windows high-contrast themes */
-	outline-color: transparent;
-	outline-width: 3px;
-	outline-style: solid;
-
-	&:active {
-		border-color: var(--ee-color-primary-high-contrast);
-	}
-}
-
-.ee-time-picker-hover {
-	border-color: var(--ee-color-primary-low-contrast);
-
-	&:active {
-		border-color: var(--ee-color-primary-high-contrast);
-	}
-}
-
-.ee-input-base-wrapper.ee-time-picker {
-	display: flex;
-	flex-flow: row;
-	justify-content: flex-start;
-	align-items: flex-start;
-
-	.react-time-picker {
-		.react-time-picker {
-			&__inputGroup {
-				border-color: var(--ee-border-color);
-				border-radius: var(--ee-border-radius-small);
-				border-style: solid;
-				border-width: 3px;
-				box-shadow: none;
-				box-sizing: border-box;
-				color: var(--ee-default-text-color);
-				line-height: inherit;
-				height: 2.625rem;
-				margin: 0 var(--ee-margin-nano) 0 0;
-				min-width: 9rem;
-				outline: none;
-				padding: var(--ee-padding-nano) var(--ee-padding-smaller);
-				transition: all ease 50ms;
-
-				&__input,
-				&__leadingZero {
-					border: none !important;
-					border-radius: unset !important;
-					box-shadow: none !important;
-					color: var(--ee-default-text-color);
-					display: inline-block;
-					font-size: var(--ee-font-size-default);
-					line-height: inherit;
-					min-height: unset !important;
-					outline: none !important;
-					margin: 0;
-					padding: unset !important;
-					transition: unset !important;
-				}
-
-				&__amPm {
-					-webkit-appearance: auto;
-					background: unset;
-					vertical-align: unset;
-
-					option {
-						background-color: var(--ee-background-color) !important;
-						color: var(--ee-default-text-color) !important;
-						height: var(--ee-icon-button-size) !important;
-					}
-
-					&:focus {
-						background-color: var(--ee-color-primary-low-contrast);
-						color: var(--ee-text-on-primary);
-						height: 22.5px;
-					}
-				}
-
-				&__divider {
-					color: var(--ee-color-grey-10);
-					font-size: var(--ee-font-size-big);
-					font-weight: 400;
-					margin: 0 var(--ee-margin-micro);
-					padding: 0;
-				}
-
-				// the following style rules do the same thing, but need to be kept separate, else they no workie
-				input[type='number']::selection {
-					background-color: var(--ee-color-primary-low-contrast);
-					color: var(--ee-text-on-primary);
-				}
-				input[type='number']::-moz-selection {
-					background-color: var(--ee-color-primary-low-contrast);
-					color: var(--ee-text-on-primary);
-				}
-				select::selection {
-					background-color: var(--ee-color-primary-low-contrast);
-					color: var(--ee-text-on-primary);
-				}
-				select::-moz-selection {
-					background-color: var(--ee-color-primary-low-contrast);
-					color: var(--ee-text-on-primary);
-				}
-
-				&:active {
-					@extend .ee-time-picker-active;
-				}
-
-				&:focus {
-					@extend .ee-time-picker-focus;
-				}
-
-				&:hover {
-					@extend .ee-time-picker-hover;
-				}
-			}
-
-			&__wrapper {
-				border: none !important;
-
-				button {
-					background: var(--ee-button-background);
-					border-color: var(--ee-button-background);
-					border-radius: var(--ee-border-radius-small);
-					border-style: solid;
-					border-width: 3px;
-					box-shadow: none;
-					color: var(--ee-button-text-color);
-					height: var(--ee-icon-button-size);
-					margin: 0 var(--ee-margin-nano);
-					padding: var(--ee-padding-micro);
-					outline: none;
-					width: var(--ee-icon-button-size);
-
-					&:active {
-						@extend .ee-time-picker-active;
-					}
-
-					&:focus {
-						@extend .ee-time-picker-focus;
-					}
-
-					&:hover {
-						@extend .ee-time-picker-hover;
-					}
-
-					svg {
-						margin: 0;
-						path {
-							fill: currentColor;
-						}
-					}
-				}
-			}
-		}
-	}
-
-	.ee-timezone-info {
-		margin: 0;
-		button {
-			margin: 0;
-			background: var(--ee-button-background);
-			border-color: var(--ee-button-background);
-
-			&:active {
-				@extend .ee-time-picker-active;
-			}
-
-			&:focus {
-				@extend .ee-time-picker-focus;
-			}
-
-			&:hover {
-				@extend .ee-time-picker-hover;
-			}
-
-			svg {
-				margin: 0;
-				path {
-					fill: currentColor;
-				}
-			}
-		}
-	}
-}
